+++ sporkfish/searcher/move_ordering/killer_move_heuristic.py
@@ -13,7 +13,7 @@ class KillerMoveHeuristic(MoveOrderHeuristic):
         killer_moves: List[List[chess.Move]],
         depth: int,
     ) -> None:
-        super().__init__()
+        MoveOrderHeuristic.__init__(self)
         self._board = board
         self._killer_moves = killer_moves
         self._depth = depth
