+++ tests/test_searcher.py
@@ -418,63 +418,3 @@ class TestNegamax:
         assert result == value
 
 
-@pytest.mark.parametrize(
-    ("fen_string", "move_scores"),
-    [
-        ("k7/8/8/8/8/8/7K/6qR w - - 1 34", [0, 50, 52]),
-    ],
-)
-class TestMvvLvaHeuristic:
-    def test_ordered_moves_end_game(
-        self, fen_string: str, move_scores: list[int]
-    ) -> None:
-        board = init_board(fen_string)
-
-        all_moves = board.legal_moves
-        num_moves = len(move_scores)
-        assert len(list(all_moves)) == num_moves
-
-        for i, move in enumerate(all_moves):
-            mo = MvvLvaHeuristic(board)
-            score = mo.evaluate(move)
-            assert score == move_scores[i]
-
-    def test_sorting_legal_moves(self, fen_string: str, move_scores: list[int]) -> None:
-        board = init_board(fen_string)
-
-        mo_heuristic = MvvLvaHeuristic(board)
-        legal_moves = MoveOrderer.order_moves(mo_heuristic, board.legal_moves)
-
-        num_moves = len(move_scores)
-        for i, move in enumerate(legal_moves):
-            score = mo_heuristic.evaluate(move)
-            assert score == move_scores[num_moves - i - 1]
-
-
-class TestKillerMoveHeuristic:
-    def test_ordered_moves_end_game(self) -> None:
-        board = init_board(board_setup["white"]["king_queen_fork"])
-
-        all_moves = board.legal_moves
-        killer_moves = [
-            [chess.Move.null(), chess.Move.null()],
-            [chess.Move.from_uci("c4b6"), chess.Move.null()],
-        ]
-        mo_heuristic = KillerMoveHeuristic(board, killer_moves, 1)
-
-        for move in enumerate(all_moves):
-            score = mo_heuristic.evaluate(move)
-            assert score == (1 if move == chess.Move.from_uci("c4b6") else 0)
-
-    def test_sorting_legal_moves(self) -> None:
-        board = init_board(board_setup["white"]["king_queen_fork"])
-        killer_moves = [
-            [chess.Move.null(), chess.Move.null()],
-            [chess.Move.from_uci("c4b6"), chess.Move.null()],
-        ]
-        mo_heuristic = KillerMoveHeuristic(board, killer_moves, 1)
-        legal_moves = MoveOrderer.order_moves(mo_heuristic, board.legal_moves)
-
-        for i, move in enumerate(legal_moves):
-            score = mo_heuristic.evaluate(move)
-            assert score == (1 if i == 0 else 0)
